Create LexboxQueryService
